# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Blindate Api",
    "description": "This is a simple API",
    "contact": {
      "email": "hafidmahdi23@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/Xyedo/blindate-api/0.0.1",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
    "/auth": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in User",
        "operationId": "postAuthHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/authReq"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successfull login",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "refreshToken=IAHH_CAHAC; maxAge:304343; path=/api/v1"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/authRes"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/resource_conflict"
          },
          "400": {
            "$ref": "#/components/responses/bad_request_json"
          },
          "401": {
            "description": "invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth_response_unauth"
                }
              }
            }
          },
          "422": {
            "description": "validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth_post_err_validation"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "when accessToken is invalid, you have to use this path to get it using your cookie refreshToken",
        "tags": [
          "Authentication"
        ],
        "operationId": "putAuthHandler",
        "summary": "to get new AccessToken",
        "security": [
          {
            "refreshAuth": []
          }
        ],
        "responses": {
          "409": {
            "$ref": "#/components/responses/resource_conflict"
          },
          "403": {
            "$ref": "#/components/responses/invalid_refresh_cookie"
          },
          "401": {
            "$ref": "#/components/responses/invalid_creds_refresh_token"
          },
          "200": {
            "description": "success refreshing accessToken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/authRes"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAuthHandler",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "refreshAuth": []
          }
        ],
        "description": "use when logging out, it removes the refreshToken",
        "summary": "using when log out",
        "responses": {
          "409": {
            "$ref": "#/components/responses/resource_conflict"
          },
          "403": {
            "$ref": "#/components/responses/invalid_refresh_cookie"
          },
          "401": {
            "$ref": "#/components/responses/invalid_creds_refresh_token"
          },
          "200": {
            "description": "success logging out",
            "content": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "refreshToken=IAHH_CAHAC; maxAge:-1; path=/api/v1"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "log out success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "bad_request_json": {
        "description": "bad request on json body",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "fail"
                },
                "message": {
                  "type": "string",
                  "example": "body contains badly-formed JSON (at character 20)"
                }
              }
            }
          }
        }
      },
      "resource_conflict": {
        "description": "conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "fail"
                },
                "message": {
                  "type": "string",
                  "example": "requirest conflicted, please try again"
                }
              }
            }
          }
        }
      },
      "invalid_refresh_cookie": {
        "description": "not having a refreshToken cookie",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/auth_response_forbidden"
            }
          }
        }
      },
      "invalid_creds_refresh_token": {
        "description": "invalid refreshToken Cookie",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/auth_response_unauth_token"
            }
          }
        }
      },
...
    },
    "schemas": {
      "authReq": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "authRes": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
      "auth_response_unauth": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "email or password is invalid"
          }
        }
      },
      "auth_response_unauth_token": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "invalid credentials"
          }
        }
      },
      "auth_response_forbidden": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Cookie not found in your browser, must be login"
          }
        }
      },
      "auth_post_err_validation": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "please refer to the documentation"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "required and must be a valid email"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "loginAuth": {
        "type": "jwt token",
        "in": "header",
        "name": "Authorization Bearer"
      },
      "refreshAuth": {
        "type": "jwt token",
        "in": "cookie",
        "name": "refreshToken"
      }
    }
  },
  "security": [
    {
      "loginAuth": []
    },
    {
      "refreshAuth": []
    }
  ]
}